Revert "[FOLIO-4239] Update Java to 21 (#171)"
This reverts commit 75a5afba17e03447528a291bf89f38e51bf09f21.
