+++ pom.xml
@@ -4,7 +4,7 @@
   <parent>
     <groupId>org.springframework.boot</groupId>
     <artifactId>spring-boot-starter-parent</artifactId>
-    <version>3.4.3</version>
+    <version>3.3.7</version>
     <relativePath /> <!-- lookup parent from repository -->
   </parent>
   <groupId>org.folio</groupId>
@@ -23,7 +23,7 @@
 
   <properties>
     <!-- runtime dependencies -->
-    <folio-spring-base.version>8.3.0-SNAPSHOT</folio-spring-base.version>
+    <folio-spring-base.version>8.2.1</folio-spring-base.version>
     <commons-collections4.version>4.4</commons-collections4.version>
     <folio-service-tools.version>3.1.0</folio-service-tools.version>
     <mapstruct.version>1.6.2</mapstruct.version>
@@ -50,7 +50,7 @@
     <maven-enforcer-plugin.version>3.1.0</maven-enforcer-plugin.version>
 
     <!-- other properties -->
-    <java.version>21</java.version>
+    <java.version>17</java.version>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
     <openapi.input.file>${project.basedir}/src/main/resources/swagger.api/list.yaml</openapi.input.file>
